requirements.txt
absl-py==1.4.0
Brotli==1.0.9
cachetools==5.3.0
certifi==2023.5.7
charset-normalizer==3.1.0
cmake==3.26.3
filelock==3.12.0
google-auth==2.17.3
google-auth-oauthlib==1.0.0
grpcio==1.54.0
idna==3.4
imageio==2.28.1
importlib-metadata==6.6.0
Jinja2==3.1.2
lazy-loader==0.2
lit==16.0.3
Markdown==3.4.3
MarkupSafe==2.1.2
mpmath==1.3.0
networkx==3.1
numpy==1.24.3
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
oauthlib==3.2.2
packaging==23.1
Pillow==9.5.0
protobuf==4.23.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pytorch-msssim==0.2.1
PyWavelets==1.4.1
PyYAML==6.0
requests==2.30.0
requests-oauthlib==1.3.1
rsa==4.9
scikit-image==0.20.0
scipy==1.9.1
six==1.16.0
sympy==1.11.1
tensorboard==2.13.0
tensorboard-data-server==0.7.0
tifffile==2023.4.12
torch==2.0.1
torchaudio==2.0.2
torchvision==0.15.2
triton==2.0.0
typing-extensions==4.5.0
urllib3==2.0.2
Werkzeug==2.3.4
zipp==3.15.0
